+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- update ts-node
+- update tslint
+- mongooseスキーマオプションを最適化
+- レポート作成処理を最適化
+- レポートのインターフェースを最適化(reservation.idを追加、aggregateUnitを削除)
+
 ### Deprecated
 
 ### Removed